refactor: 공통 로직 분리 및 궁성에서 특정 경로가 허용되는 기물에 대한 추상화 추가
