+++ src/main/java/janggi/piece/Cha.java
@@ -3,21 +3,9 @@ package janggi.piece;
 import janggi.position.Path;
 import janggi.position.Position;
 import janggi.team.TeamType;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 
-public class Cha extends Piece {
-
-    private static final List<Integer> FORWARD_RIGHT_DIAGONAL = List.of(1, 1);
-    private static final List<Integer> FORWARD_LEFT_DIAGONAL = List.of(1, -1);
-    private static final List<Integer> BACKWARD_RIGHT_DIAGONAL = List.of(-1, 1);
-    private static final List<Integer> BACKWARD_LEFT_DIAGONAL = List.of(-1, -1);
-
-    private static final Set<List<Integer>> AVAILABLE_DIFFERENCE_IN_GUNGSUNG = Set.of(
-            FORWARD_RIGHT_DIAGONAL, FORWARD_LEFT_DIAGONAL,
-            BACKWARD_RIGHT_DIAGONAL, BACKWARD_LEFT_DIAGONAL
-    );
+public class Cha extends GungSungPiece {
 
     public Cha(TeamType teamType) {
         super(PieceType.CHA, teamType);
@@ -25,17 +13,7 @@ public class Cha extends Piece {
 
     @Override
     public Path makePath(Position currentPosition, Position arrivalPosition) {
-        int differenceForY = arrivalPosition.calculateDifferenceForY(currentPosition);
-        int differenceForX = arrivalPosition.calculateDifferenceForX(currentPosition);
-
-        if (isMovingInOnlyGungSung(currentPosition, arrivalPosition) && isDiagonalInGungSung(currentPosition,
-                differenceForY, differenceForX)) {
-            return new Path(
-                    calculateGungSungMovingPositions(currentPosition, arrivalPosition, differenceForY, differenceForX));
-        }
-
-        validateDistanceAndDirection(differenceForY, differenceForX);
-        return new Path(calculateMovingPositions(currentPosition, arrivalPosition, differenceForY, differenceForX));
+        return makePathForGungSungPiece(currentPosition, arrivalPosition);
     }
 
     @Override
@@ -45,40 +23,15 @@ public class Cha extends Piece {
         }
     }
 
-    List<Position> calculateGungSungMovingPositions(Position currentPosition, Position arrivalPosition,
-                                                    int differenceForY,
-                                                    int differenceForX) {
-        final List<Position> positions = new ArrayList<>();
-        int currentY = currentPosition.getY();
-        int currentX = currentPosition.getX();
-
-        calculatePath(arrivalPosition, differenceForY, differenceForX, positions, currentY, currentX);
-        return positions;
-    }
-
-    private void calculatePath(Position arrivalPosition, int differenceForY, int differenceForX,
-                               List<Position> positions, int currentY, int currentX) {
-        int differenceUnitY = calculateUnit(differenceForY);
-        int differenceUnitX = calculateUnit(differenceForX);
-        while (currentX != arrivalPosition.getX() && currentY != arrivalPosition.getY()) {
-            currentY += differenceUnitY;
-            currentX += differenceUnitX;
-            positions.add(Position.valueOf(currentY, currentX));
-        }
+    @Override
+    boolean isDiagonalInGungSung(Position currentPosition, int differenceForY, int differenceForX) {
+        return Position.isAbleToDiagonalMoveInGungSung(currentPosition) && AVAILABLE_DIFFERENCE_IN_GUNGSUNG.contains(
+                List.of(calculateUnit(differenceForY), calculateUnit(differenceForX)));
     }
 
     private boolean canNotMove(int differenceForY, int differenceForX) {
         return !((Math.abs(differenceForY) > 0 && Math.abs(differenceForX) == 0) ||
                 (Math.abs(differenceForY) == 0 && Math.abs(differenceForX) > 0));
-    }
-
-    private boolean isMovingInOnlyGungSung(Position currentPosition, Position arrivalPosition) {
-        return Position.isInGungSung(currentPosition) && Position.isInGungSung(arrivalPosition);
-    }
-
-    private boolean isDiagonalInGungSung(Position currentPosition, int differenceForY, int differenceForX) {
-        return Position.isAbleToDiagonalMoveInGungSung(currentPosition) && AVAILABLE_DIFFERENCE_IN_GUNGSUNG.contains(
-                List.of(calculateUnit(differenceForY), calculateUnit(differenceForX)));
     }
 }
