+++ src/main/java/janggi/piece/Ma.java
@@ -1,6 +1,5 @@
 package janggi.piece;
 
-import janggi.position.Path;
 import janggi.position.Position;
 import janggi.team.TeamType;
 import java.util.List;
@@ -26,16 +25,6 @@ public class Ma extends Piece {
 
     public Ma(TeamType teamType) {
         super(PieceType.MA, teamType);
-    }
-
-    @Override
-    public Path makePath(Position currentPosition, Position arrivalPosition) {
-        int differenceForY = arrivalPosition.calculateDifferenceForY(currentPosition);
-        int differenceForX = arrivalPosition.calculateDifferenceForX(currentPosition);
-
-        validateDistanceAndDirection(differenceForY, differenceForX);
-
-        return new Path(calculateMovingPositions(currentPosition, arrivalPosition, differenceForY, differenceForX));
     }
 
     @Override
