+++ src/main/java/janggi/piece/Po.java
@@ -3,11 +3,10 @@ package janggi.piece;
 import janggi.position.Path;
 import janggi.position.Position;
 import janggi.team.TeamType;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
-public class Po extends Piece {
+public class Po extends GungSungPiece {
 
     private static final List<Integer> FORWARD_RIGHT_DIAGONAL = List.of(1, 1);
     private static final List<Integer> FORWARD_LEFT_DIAGONAL = List.of(1, -1);
@@ -25,17 +24,7 @@ public class Po extends Piece {
 
     @Override
     public Path makePath(Position currentPosition, Position arrivalPosition) {
-        int differenceForY = arrivalPosition.calculateDifferenceForY(currentPosition);
-        int differenceForX = arrivalPosition.calculateDifferenceForX(currentPosition);
-
-        if (isMovingInOnlyGungSung(currentPosition, arrivalPosition) && isDiagonalInGungSung(currentPosition,
-                differenceForY, differenceForX)) {
-            return new Path(
-                    calculateGungSungMovingPositions(currentPosition, arrivalPosition, differenceForY, differenceForX));
-        }
-
-        validateDistanceAndDirection(differenceForY, differenceForX);
-        return new Path(calculateMovingPositions(currentPosition, arrivalPosition, differenceForY, differenceForX));
+        return makePathForGungSungPiece(currentPosition, arrivalPosition);
     }
 
     @Override
@@ -43,37 +32,6 @@ public class Po extends Piece {
         if (isNotAbleToMoveDirection(differenceForY, differenceForX)) {
             throw new IllegalArgumentException("[ERROR] 포는 이어진 선을 따라서만 이동할 수 있습니다.");
         }
-    }
-
-    List<Position> calculateGungSungMovingPositions(Position currentPosition, Position arrivalPosition,
-                                                    int differenceForY,
-                                                    int differenceForX) {
-        final List<Position> positions = new ArrayList<>();
-        int currentY = currentPosition.getY();
-        int currentX = currentPosition.getX();
-
-        calculatePath(arrivalPosition, differenceForY, differenceForX, positions, currentY, currentX);
-        return positions;
-    }
-
-    private void calculatePath(Position arrivalPosition, int differenceForY, int differenceForX,
-                               List<Position> positions, int currentY, int currentX) {
-        int differenceUnitY = calculateUnit(differenceForY);
-        int differenceUnitX = calculateUnit(differenceForX);
-        while (currentX != arrivalPosition.getX() && currentY != arrivalPosition.getY()) {
-            currentY += differenceUnitY;
-            currentX += differenceUnitX;
-            positions.add(Position.valueOf(currentY, currentX));
-        }
-    }
-
-    private boolean isMovingInOnlyGungSung(Position currentPosition, Position arrivalPosition) {
-        return Position.isInGungSung(currentPosition) && Position.isInGungSung(arrivalPosition);
-    }
-
-    private boolean isDiagonalInGungSung(Position currentPosition, int differenceForY, int differenceForX) {
-        return Position.isAbleToDiagonalMoveInGungSung(currentPosition) && AVAILABLE_DIFFERENCE_IN_GUNGSUNG.contains(
-                List.of(calculateUnit(differenceForY), calculateUnit(differenceForX)));
     }
 
     @Override
@@ -85,6 +43,13 @@ public class Po extends Piece {
             throw new IllegalArgumentException("[ERROR] 포는 포끼리 뛰어넘거나 잡을 수 없습니다.");
         }
     }
+    
+    @Override
+    boolean isDiagonalInGungSung(Position currentPosition, int differenceForY, int differenceForX) {
+        return Position.isAbleToDiagonalMoveInGungSung(currentPosition) && AVAILABLE_DIFFERENCE_IN_GUNGSUNG.contains(
+                List.of(calculateUnit(differenceForY), calculateUnit(differenceForX)));
+    }
+
 
     private boolean isNotAbleToMoveDirection(int differenceForY, int differenceForX) {
         return !((Math.abs(differenceForY) > 0 && Math.abs(differenceForX) == 0) ||
